use correct backend and darkmode fixes

+++ DP3TApp/Logic/Statistics/StatisticsResponse.swift
@@ -17,7 +17,7 @@ class StatisticsResponse: Codable {
     class StatisticEntry: Codable {
         let date: Date
         let newInfections: Int?
-        let newInfectionsSevenDayAverage: Double?
+        let newInfectionsSevenDayAverage: Int?
         let covidcodesEntered: Int?
     }
 }
